log: Tie log_level to opt.loglevel
Gives some very nice consistency features especially during early startup

Note: It breaks the rule of letting opts always reflect only invocation
intent. For such an important feature as logging we can make an exception.
It does however also put a inwards dependency, which is less nice as log is
intended to be a self-contained module. We might have to have a closer look
at that.

Loglevel/verbosity is more of a debug aid than functionality, therefor we're
less stringent with principles.