bin/cont_trace.sh: Somehow managed to miss a event
Missing event: sched_wakeup_new

Because probably of this statistics didn't add up when making a sharp run
where:

* Out of every even not containing "sched_stat"
* Number of sched_switch events should be *exactly half* of the remaining
  number of events.

This event indicates however newly created threads which is unfortunate as
it shows this process might create (and destroy) threads dynamically. (Or it
could be forks, which in such case is less relevant.)

The missing amount of events is quite low however. In a 10 seconds long
Google-chrome probing under heavy load (sched_stat not accounted for) it
was: 1605/181017, i.e. 0.8%