Option added: -s --buffer-size
Optionally set buffer size prior run. If option not seen, don't touch and
run with whatever is there.

This option was added hoping to solve non-monotonic time-stamping but even
with 99MB buffer size per second solves the issue. It could be a combination
error as now duplicates can be seen.

Testing:

$ source $SOMEDIR/bin/cont_trace.sh

$ sudo  ./etrace -t \
	-e sched/sched_switch -f '(prev_pid == %tid%) || (next_pid == %tid%)' \
	-o- -verror -ccounter -s99000 -T1000000 -p$(maxof chrome) > $ETRACE_FILE

$ to_sampler.sh $ETRACE_FILE $(
	pids.sh /tmp/cont_etrace2.etrace | \
 	awk '{printf("%s,",$1);}END{printf("\n");}' | \
 	sed -e 's/.$//') > $FILE_SPRAW

$ for ((N=0; N<7; N++)); do \
	cat $FILE_SPRAW | \
		sort -n -k1 -t";" | \
		awk -F";" '$3=='$N'{
			if ((L6==$6) && (L7==$7) && ($6!=$7)){print};
			L6=$6;L7=$7;
		}';
	done

There should be no prints if output is OK.